Always process DNS and DHCP settings in installer (#1630)
* Always process DNS and DHCP settings in installer

* Make sure dnsmasq config exists before modifying it


* Make sure the dnsmasq config directory exists


* Only remove the DHCP config if it exists (fixes tests, hopefully)

+++ advanced/Scripts/webpage.sh
@@ -292,7 +292,9 @@ ra-param=*,0,0
     fi
 
 	else
-		rm "${dhcpconfig}" &> /dev/null
+	  if [[ -f "${dhcpconfig}" ]]; then
+		  rm "${dhcpconfig}" &> /dev/null
+		fi
 	fi
 }
